Use github.com/npat-efault/poller to read input events
Use github.com/npat-efault/poller to read input events instead of the Go "os"
package. This enables off-loading event reading to separate goroutines when one
also wishes to be able to shutdown these goroutines. When using the "os" package
for file I/O it's not always possible to signal shutdown to a goroutine blocked
on a read operation by closing the file. The "poller" package enables this.